Allow arbitrary selection of backup sources
At the moment, our subversion backup is the largest part of our
backup.  Our subversion is read-only and has been for years.
Therefore, we don't really need to back it up more than once...

I think the normal use-case of backup.py will be to ask it for a
backup of everything except for svn.  Previously, the way to do this
was to invoke backup.py separately for every backup source except for
svn.  This may mean that if we add a backup source in the future, we
might inadvertently miss out on backing it up.

This patch makes making such backups much more convenient.  backup.py
now takes multiple backup sources as arguments, e.g:

	./backup.py ide gerrit

As before, one can backup absolutely everything by specifying 'all':

	./backup.py all

One can also exclude things from the set of things one wants to
backup by prefixing with a '-':

	./backup.py -- all -svn

(Note that the '--' is a standard way of indicating to the argument
parser that everything to the right is not a flag.)

Note that it is evaluated from left-to-right, not as a set of
constraints.  So this command is not sensible:

	./backup.py -- -svn all

Change-Id: Iaed94f4601593e0d2db1247a065d284e8d161ade