Remove optional nature of three files.
This optional situation exists because I was running backup.py on
optimus to move data over to badger; naturally optimus didn't have all
the puppet gunge installed, so the backup kept on failing.

From now on we're going to be using puppet everywhere; so don't allow
these files to be optional. If they're not there, it's a backup failure.