Fix restoration of git-admin group
This is probably defunct, but anyway; we back it up, and on restoration
the backup in the ldap dump conflicts with the puppet-created group. So,
add it to the list of things that should be restored by modifying
existing groups.

Do the same for 'withdrawn', that's data we want to preserve. Remove
'mentors' from list of groups to delete, due to the control flow
structure it's never deleted (and if it were, that would be bad).

Change-Id: I6ae6552d9f382f57dcf8e681331a8549eed669de